requirements.txt
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
click==8.0.3
fastapi==0.73.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
pip==22.0.3
pluggy==1.0.0
py==1.11.0
pydantic==1.9.0
pyparsing==3.0.7
pytest==7.0.1
setuptools==60.6.0
sniffio==1.2.0
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.1.1
uvicorn==0.17.5
wheel==0.37.1
torch
torchvision
supervision
Pillow
mysql-connector-python